patch(container): sync with upstream
This patch sync's with upstreams latest release to fix a race condition. `S6_SERVICES_READYTIME` 
 (default = 5): With services declared in `/etc/services.d`, there is
an unavoidable race condition between the moment when services are started and the moment when
they can be tested for readiness. To avoid that race, we sleep a little time, by default 5
milliseconds, before testing for readiness. 

Source commit:
https://github.com/just-containers/s6-overlay/commit/94ba8e04c2227720b5941bf6a84920cc464f1b12